+++ my-app/src/tests/pages/Delivery/DeliverySpreadsheet.styles.test.ts
@@ -6,7 +6,7 @@ describe("DeliverySpreadsheet styles regression", () => {
   // Guards against reintroducing global html/body centering styles that can shift
   // page flow and hide or misplace route search/map sections.
   it("does not apply global html/body centering styles", () => {
-    const cssPath = path.resolve(__dirname, "DeliverySpreadsheet.css");
+    const cssPath = path.resolve(__dirname, "../../../pages/Delivery/DeliverySpreadsheet.css");
     const css = fs.readFileSync(cssPath, "utf8");
 
     expect(css).not.toMatch(/(^|\n)\s*html\s*,\s*\n\s*body\s*\{/m);
