+++ my-app/src/tests/pages/Delivery/utils/routeAssignmentState.test.ts
@@ -9,7 +9,7 @@ import {
   renumberRoutesSequentially,
   updateClientRouteAssignment,
   type RouteAssignmentState,
-} from "./routeAssignmentState";
+} from "../../../../pages/Delivery/utils/routeAssignmentState";
 
 const createState = (): RouteAssignmentState => ({
   clusters: [
